requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1
SQLAlchemy==2.0.23

# Redis
redis[hiredis]==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# SSH and PTY support
paramiko==3.3.1
pexpect==4.8.0

# HTTP client for AI services
httpx==0.25.2

# Configuration
python-dotenv==1.0.0

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.1

# CORS
fastapi-cors==0.0.6

# Testing - Core
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2

# Testing - Enhanced
pytest-cov==4.1.0           # Coverage reporting
pytest-xdist==3.5.0         # Parallel test execution
pytest-html==4.1.1          # HTML test reports
factory-boy==3.3.0          # Test data factories
faker==20.1.0               # Realistic test data
pytest-benchmark==4.0.0     # Performance testing
pytest-timeout==2.1.0       # Test timeout handling
pytest-env==1.1.1           # Environment variable handling
pytest-randomly==3.15.0     # Random test execution order

# Development tools
black==23.11.0
ruff==0.1.6
mypy==1.7.1

# Logging
structlog==23.2.0

# Monitoring
prometheus-client==0.19.0